Custom Orders functionality

+++ app/src/main/java/com/nextnut/logistica/CustomDetailActivity.java
@@ -1,18 +1,42 @@
 package com.nextnut.logistica;
 
+import android.animation.Animator;
+import android.animation.AnimatorListenerAdapter;
+import android.annotation.TargetApi;
+import android.app.LoaderManager;
+import android.content.CursorLoader;
 import android.content.Intent;
+import android.content.Loader;
+import android.content.pm.PackageManager;
+import android.database.Cursor;
+import android.net.Uri;
+import android.os.AsyncTask;
+import android.os.Build;
 import android.os.Bundle;
+import android.provider.ContactsContract;
+import android.support.annotation.NonNull;
 import android.support.design.widget.FloatingActionButton;
 import android.support.design.widget.Snackbar;
 import android.support.v7.widget.Toolbar;
+import android.text.TextUtils;
 import android.util.Log;
+import android.view.KeyEvent;
 import android.view.View;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.app.ActionBar;
 import android.support.v4.app.NavUtils;
 import android.view.MenuItem;
+import android.view.inputmethod.EditorInfo;
+import android.widget.ArrayAdapter;
+import android.widget.AutoCompleteTextView;
+import android.widget.Button;
+import android.widget.EditText;
+import android.widget.TextView;
 
-import com.nextnut.logistica.data.CustomColumns;
+import java.util.ArrayList;
+import java.util.List;
+
+import static android.Manifest.permission.READ_CONTACTS;
 
 /**
  * An activity representing a single Custom detail screen. This
@@ -89,7 +113,7 @@ public class CustomDetailActivity extends AppCompatActivity {
             // Create the detail fragment and add it to the activity
             // using a fragment transaction.
             Bundle arguments = new Bundle();
-            arguments.putInt(CustomDetailFragment.ARG_ITEM_ID, getIntent().getIntExtra(CustomDetailFragment.ARG_ITEM_ID,0));
+            arguments.putLong(CustomDetailFragment.ARG_ITEM_ID, getIntent().getLongExtra(CustomDetailFragment.ARG_ITEM_ID,0));
             mAction= getIntent().getIntExtra(CustomDetailFragment.CUSTOM_ACTION,CustomDetailFragment.CUSTOM_SELECTION);
             arguments.putInt(ProductDetailFragment.PRODUCT_ACTION,mAction);
 
@@ -131,4 +155,6 @@ public class CustomDetailActivity extends AppCompatActivity {
         }
         return super.onOptionsItemSelected(item);
     }
+
+
 }
